refactor(project): split code build script to in preparation of type build script
